CCO Update

+++ Other/Another Contest 7 Problem 1 - Lonely Users.cpp
@@ -14,7 +14,7 @@ typedef  unsigned long long ull;
 typedef long long ll;
 typedef long double ld;
 using namespace std; 
-int T, N; 
+ll T, N; 
 int main(){
 	cin>>T; 
 	while(T--){
